suppress CA1416(PlatformNotSupport) warning
